check human before confirming call and style get responses

+++ server/routes/getCall.js
@@ -5,13 +5,15 @@ const { VoiceResponse } = require('twilio').twiml;
 const Provider = require('../models/Provider');
 const Call = require('../models/Call');
 const getDate = require('../modules/getDate');
+const formatHTML = require('../modules/formatHTML');
 
-const { NEBOTS_TWACCOUNTSID, NEBOTS_TWAUTHTOKEN } = process.env;
+const { NEBOTS_TWACCOUNTSID, NEBOTS_TWAUTHTOKEN, NEBOTS_SERVER } = process.env;
 
 const tw = require('twilio')(NEBOTS_TWACCOUNTSID, NEBOTS_TWAUTHTOKEN);
 
 module.exports = async (req, res) => {
   try {
+    const confirm = req.query.confirm;
     const provider = await Provider.findOne({
       _id: req.query.provider
     }).exec();
@@ -24,35 +26,50 @@ module.exports = async (req, res) => {
       console.error(`${getDate()} - Call not found`);
       return res
         .status(200)
-        .send('<h1>Gràcies, la trucada ja ha estat atesa</h1>');
+        .send(formatHTML('Gràcies, la trucada ja ha estat atesa'));
     }
 
     if (!provider) {
       console.error(`${getDate()} - Provider not found`);
-      return res.status(400).send('<h1>Alguna cosa ha anat malament</h1>');
+      return res.status(400).send(formatHTML('Alguna cosa ha anat malament'));
     }
 
-    const response = new VoiceResponse();
+    if (confirm) {
+      const response = new VoiceResponse();
 
-    response.dial(provider.phone);
+      response.dial(provider.phone);
 
-    tw.calls(call.callSid).update({ twiml: response.toString() });
+      tw.calls(call.callSid).update({ twiml: response.toString() });
 
-    Call.deleteOne({ _id: req.params.id }, () => {});
+      Call.deleteOne({ _id: req.params.id }, () => {});
 
-    console.log(`${getDate()} - Establishing call`);
+      console.log(`${getDate()} - Establishing call`);
 
-    return res
-      .status(200)
-      .send(
-        `<h1>Gràcies. Si l'usuari segueix en línia iniciarem la connexió</h1>`
+      return res
+        .status(200)
+        .send(
+          formatHTML(
+            `Gràcies. Si l'usuari segueix en línia iniciarem la connexió`
+          )
+        );
+    } else {
+      console.log(`${getDate()} - Confirming call`);
+      return res.status(200).send(
+        formatHTML(
+          `<a style="color:#003146" href="${NEBOTS_SERVER}/call/${call._id}?provider=${provider._id}&confirm=true">
+            Fes click aquí per confirmar la trucada
+            </a>`
+        )
       );
+    }
   } catch (error) {
     console.error(`${getDate()} - Error establishing call: ${error.message}`);
     return res
       .status(500)
       .send(
-        '<h1>Connexió no establerta. Pot ser que ja no calgui la teva ajuda. Gràcies!</h1>'
+        formatHTML(
+          'Connexió no establerta. Pot ser que ja no calgui la teva ajuda. Gràcies!'
+        )
       );
   }
 };
